Changes function signatures that deal with timestamp to always accept or return int64 values. Partial fix of #2677.

+++ src/apps/chifra/internal/list/handle_list.go
@@ -53,8 +53,8 @@ func (opts *ListOptions) HandleListAppearances(monitorArray []monitor.Monitor) e
 						Address:          mon.Address,
 						BlockNumber:      app.BlockNumber,
 						TransactionIndex: app.TransactionId,
-						Timestamp:        int64(ts),
-						Date:             gostradamus.FromUnixTimestamp(int64(ts)).String(),
+						Timestamp:        ts,
+						Date:             gostradamus.FromUnixTimestamp(ts).String(),
 					}
 					if uint64(record+1) >= opts.FirstRecord && (opts.MaxRecords == 250 || uint64(len(verboseResults)) < opts.MaxRecords) {
 						verboseResults = append(verboseResults, s)
