+++ src/apps/chifra/pkg/identifiers/resolve.go
@@ -82,7 +82,7 @@ func snapBnToPeriod(bn uint64, chain, period string) (uint64, error) {
 	if dt.Time().Before(firstDate.Time()) {
 		dt = firstDate
 	}
-	ts := uint64(dt.UnixTimestamp())
+	ts := dt.UnixTimestamp()
 	return tslib.FromTsToBn(chain, ts)
 }
 
@@ -124,7 +124,7 @@ func (id *Identifier) nextBlock(chain string, current uint64) (uint64, error) {
 				// should not happen
 			}
 
-			ts := uint64(dt.UnixTimestamp())
+			ts := dt.UnixTimestamp()
 			bn, err = tslib.FromTsToBn(chain, ts)
 			if err != nil {
 				return bn, err
@@ -149,12 +149,12 @@ func (p *Point) resolvePoint(chain string) uint64 {
 		bn, _ = tslib.FromNameToBn(chain, p.Special)
 	} else if p.Number >= utils.EarliestEvmTs {
 		var err error
-		bn, err = tslib.FromTsToBn(chain, uint64(p.Number))
+		bn, err = tslib.FromTsToBn(chain, int64(p.Number))
 		if err == tslib.ErrInTheFuture {
 			provider := config.GetRpcProvider(chain)
 			latest := rpcClient.BlockNumber(provider)
 			tsFuture := rpc.GetBlockTimestamp(chain, latest)
-			secs := (uint64(tsFuture) - uint64(p.Number))
+			secs := uint64(tsFuture - int64(p.Number))
 			blks := (secs / 13)
 			bn = latest + blks
 		}
